Formatting: Return a more precise list of TextEdits
Previously, the ReScript language server responded to
`textDocument/formatting` requests with a single TextEdit containing the
entire file, even for a change affecting a single line.

With this change, the file's contents before and after formatting are
diffed, and, based on the diff, a list of TextEdits sufficient to
describe the changes are returned.

The diff implementation is adapted[1] from ocaml-lsp and is based on the
simplediff algorithm[2].

[1]: https://github.com/ocaml/ocaml-lsp/pull/434
[2]: https://github.com/paulgb/simplediff